Expire access_token and id_token cookies
Currently the access_token and id_token cookies are session cookies, so
they expire in the browser when the current browser session ends.
However, the tokens have a limited lifetime on the identity provider,
so sending the cookies past their lifetime leads to an unauthorized
error.

This change makes the access_token and id_token cookies expire when
they expire on the identity provider. This leads to a better user
experience.